+++ webots_ros2_driver/src/plugins/static/Ros2Lidar.cpp
@@ -14,8 +14,8 @@
 
 #include <webots_ros2_driver/plugins/static/Ros2Lidar.hpp>
 
-#include <sensor_msgs/msg/point_field.hpp>
 #include <geometry_msgs/msg/transform_stamped.hpp>
+#include <sensor_msgs/msg/point_field.hpp>
 
 #include <webots/robot.h>
 
@@ -33,8 +33,7 @@ namespace webots_ros2_driver
     mIsPointCloudEnabled = false;
 
     // Laser publisher
-    if (wb_lidar_get_number_of_layers(mLidar) == 1)
-    {
+    if (wb_lidar_get_number_of_layers(mLidar) == 1) {
       mLaserPublisher = mNode->create_publisher<sensor_msgs::msg::LaserScan>(mTopicName, rclcpp::SensorDataQoS().reliable());
       const int resolution = wb_lidar_get_horizontal_resolution(mLidar);
       mLaserMessage.header.frame_id = mFrameName;
@@ -49,7 +48,8 @@ namespace webots_ros2_driver
     }
 
     // Point cloud publisher
-    mPointCloudPublisher = mNode->create_publisher<sensor_msgs::msg::PointCloud2>(mTopicName + "/point_cloud", rclcpp::SensorDataQoS().reliable());
+    mPointCloudPublisher =
+      mNode->create_publisher<sensor_msgs::msg::PointCloud2>(mTopicName + "/point_cloud", rclcpp::SensorDataQoS().reliable());
     mPointCloudMessage.header.frame_id = mFrameName;
     mPointCloudMessage.height = 1;
     mPointCloudMessage.point_step = 20;
@@ -77,8 +77,7 @@ namespace webots_ros2_driver
     }
   }
 
-  void Ros2Lidar::step()
-  {
+  void Ros2Lidar::step() {
     if (!preStep())
       return;
 
@@ -92,12 +91,11 @@ namespace webots_ros2_driver
       return;
 
     const bool shouldPointCloudBeEnabled = mPointCloudPublisher->get_subscription_count() > 0;
-    const bool shouldSensorBeEnabled = shouldPointCloudBeEnabled ||
-                                 (mLaserPublisher != nullptr && mLaserPublisher->get_subscription_count() > 0);
+    const bool shouldSensorBeEnabled =
+      shouldPointCloudBeEnabled || (mLaserPublisher != nullptr && mLaserPublisher->get_subscription_count() > 0);
 
     // Enable/Disable sensor
-    if (shouldSensorBeEnabled != mIsSensorEnabled)
-    {
+    if (shouldSensorBeEnabled != mIsSensorEnabled) {
       if (shouldSensorBeEnabled)
         wb_lidar_enable(mLidar, mPublishTimestepSyncedMs);
       else
@@ -106,8 +104,7 @@ namespace webots_ros2_driver
     }
 
     // Enable/Disable point cloud
-    if (shouldPointCloudBeEnabled != mIsPointCloudEnabled)
-    {
+    if (shouldPointCloudBeEnabled != mIsPointCloudEnabled) {
       if (shouldPointCloudBeEnabled)
         wb_lidar_enable_point_cloud(mLidar);
       else
@@ -116,11 +113,9 @@ namespace webots_ros2_driver
     }
   }
 
-  void Ros2Lidar::publishPointCloud()
-  {
+  void Ros2Lidar::publishPointCloud() {
     auto data = wb_lidar_get_point_cloud(mLidar);
-    if (data)
-    {
+    if (data) {
       mPointCloudMessage.header.stamp = mNode->get_clock()->now();
 
       mPointCloudMessage.width = wb_lidar_get_number_of_points(mLidar);
@@ -133,15 +128,13 @@ namespace webots_ros2_driver
     }
   }
 
-  void Ros2Lidar::publishLaserScan()
-  {
+  void Ros2Lidar::publishLaserScan() {
     auto rangeImage = wb_lidar_get_layer_range_image(mLidar, 0);
-    if (rangeImage)
-    {
+    if (rangeImage) {
       memcpy(mLaserMessage.ranges.data(), rangeImage, mLaserMessage.ranges.size() * sizeof(float));
       mLaserMessage.header.stamp = mNode->get_clock()->now();
       mLaserPublisher->publish(mLaserMessage);
     }
   }
 
-} // end namespace webots_ros2_driver
+}  // end namespace webots_ros2_driver