+++ webots_ros2/CHANGELOG.rst
@@ -2,9 +2,29 @@
 Changelog for package webots_ros2
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+2023.0.3 (2023-XX-XX)
+------------------
+* Improved the navigation of the TIAGo example.
+* Replaced Slam-Toolbox by Cartographer in the TIAGo example.
+* Adding port, stream type parameters to webots_laucher
+* Copying .wbproj when launching a Webots world via webots_launcher
+* Added Emitter and Receiver support in webots_ros2_driver
+
+2023.0.2 (2023-02-07)
+------------------
+* Drop support for Galactic.
+* Fixed the spawn of URDF robots in WSL and macOS when using full path.
+* Fixed relative assets in macOS.
+* Ros2Supervisor is now optional.
+
+2023.0.1 (2023-01-05)
+------------------
+* Fixed relative assets in WSL.
+* Fixed broken controller connection in Rats life example.
+
 2023.0.0 (2022-11-30)
 ------------------
-* Add support for the new Python API of Webots R2023a
+* Added support for the new Python API of Webots R2023a
 * Convert C++ controller API functions to C
 * Replace libController submodule by commited source files
 * Removed 'webots_ros2_core' package (deprecated).
@@ -12,7 +32,7 @@ Changelog for package webots_ros2
 
 2022.1.4 (2022-11-18)
 ------------------
-* Fix the camera focal length in the CameraInfo topic.
+* Fixed the camera focal length in the CameraInfo topic.
 * Upgraded to urdf2webots 2.0.3
 * Update the calculation of CameraRecognitionObject messages to the RDF convention of R2022b.