+++ webots_ros2_epuck/CHANGELOG.rst
@@ -2,6 +2,14 @@
 Changelog for package webots_ros2_epuck
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+2023.0.2 (2023-02-07)
+------------------
+* Updated supervisor launch.
+
+2023.0.1 (2023-01-05)
+------------------
+* Fixed broken controller connection in Rats life example.
+
 2022.1.3 (2022-11-02)
 ------------------
 * Added macOS support.